Move retrieveNonce to the PoolState module
