--- packages/hardhat-core/src/internal/hardhat-network/provider/utils/retrieveNonce.ts
@@ -1,7 +1,0 @@
-import { BN, toBuffer } from "ethereumjs-util";
-
-import { SerializedTransaction } from "../PoolState";
-
-export function retrieveNonce(tx: SerializedTransaction) {
-  return new BN(toBuffer(tx.get("data").get(0)));
-}
