+++ packages/hardhat-core/test/internal/hardhat-network/provider/retrieveNonce.ts
@@ -1,7 +1,7 @@
 import { assert } from "chai";
 
-import { retrieveNonce } from "../../../../../src/internal/hardhat-network/provider/utils/retrieveNonce";
-import { createTestSerializedTransaction } from "../../helpers/blockchain";
+import { retrieveNonce } from "../../../../src/internal/hardhat-network/provider/PoolState";
+import { createTestSerializedTransaction } from "../helpers/blockchain";
 
 describe("retrieveNonce", () => {
   it("can retrieve nonce from serialized transaction", () => {
